Reorganize data and components to add a 404 page
