+++ .gitignore
@@ -3,5 +3,6 @@ node_modules
 coverage
 dist
 .next
+.data.json
 *.tsbuildinfo
 *.log
