+++ packages/site/next.config.mjs
@@ -1,7 +1,29 @@
-export default {
-  cleanUrls: true,
-  trailingSlash: false,
-  eslint: {
-    ignoreDuringBuilds: true
+import { createRequire } from "module"
+import path from "path"
+import { writeJsonFileSync } from "write-json-file"
+
+const DATA_PATH = path.resolve("./src/.data.json")
+
+/**
+ * Create a static data object.
+ */
+function getData() {
+  const pkg = createRequire(import.meta.url)("./package.json")
+
+  return {
+    version: pkg.version,
+    date: String(new Date().getFullYear())
+  }
+}
+
+export default () => {
+  writeJsonFileSync(DATA_PATH, getData())
+
+  return {
+    cleanUrls: true,
+    trailingSlash: false,
+    eslint: {
+      ignoreDuringBuilds: true
+    }
   }
 }
