+++ packages/site/package.json
@@ -68,6 +68,7 @@
     "postcss-nesting": "^10.1.0",
     "stylelint": "^14.2.0",
     "tailwindcss": "^3.0.7",
-    "type-fest": "^2.8.0"
+    "type-fest": "^2.8.0",
+    "write-json-file": "^5.0.0"
   }
 }
